Finalização API
Finalização dos métodos referentes aos acessos da API.

+++ huapi/src/main/java/br/com/huapi/repository/ContatoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.huapi.domain.Contato;
 
-public interface ContatoRepository extends JpaRepository<Contato, Integer>{
+public interface ContatoRepository extends JpaRepository<Contato, Long>{
 
 }
