+++ huapi/src/main/java/br/com/huapi/repository/FuncionarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.huapi.domain.Funcionario;
 
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Integer>{
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Long>{
 
 }
