+++ huapi/src/main/java/br/com/huapi/repository/GuiaAtendimentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.huapi.domain.GuiaAtendimento;
 
-public interface GuiaAtendimentoRepository extends JpaRepository<GuiaAtendimento, Integer> {
+public interface GuiaAtendimentoRepository extends JpaRepository<GuiaAtendimento, Long> {
 
 }
