+++ huapi/src/main/java/br/com/huapi/repository/HospitalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.huapi.domain.Hospital;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer>{
+public interface HospitalRepository extends JpaRepository<Hospital, Long>{
 
 }
